+++ mtpy/modeling/ModEM/__init__.py
@@ -12,9 +12,10 @@ from .plot_response import PlotResponse
 from .plot_pt_maps import PlotPTMaps
 from .plot_depth_slice import PlotDepthSlice
 from .plot_slices import PlotSlices
+from .plot_rms_maps import PlotRMSMaps
 
 __all__ = ['ModEMError', 'DataError', 'Stations', 'Data', 'Model', 'Residual',
            'ControlInv', 'ControlFwd', 'Covariance', 'ModEMConfig', 'ModelManipulator',
-           'PlotResponse', 'PlotPTMaps', 'PlotDepthSlice', 'PlotSlices']
+           'PlotResponse', 'PlotPTMaps', 'PlotDepthSlice', 'PlotSlices', 'PlotRMSMaps']
